CARDS-2333: Automatically check in forms that haven't been modified in more than 30 minutes Move the autocheckin code in the cards-versioning module
