+++ modules/versioning/src/main/features/feature.json
@@ -1,0 +1,38 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+{
+    "bundles":[
+        {
+            "id":"io.uhndata.cards:cards-versioning:${project.version}",
+            "start-order":"25"
+        }
+    ],
+    "configurations":{
+        "org.apache.sling.jcr.repoinit.RepositoryInitializer~cards-versioning":{
+            "service.ranking:Integer":150,
+            "scripts":[
+                "create service user autocheckin \n set ACL on /Forms \n   allow jcr:all for autocheckin \n end"
+            ]
+        },
+        "org.apache.sling.serviceusermapping.impl.ServiceUserMapperImpl.amended~cards-versioning":{
+            "user.mapping":[
+                "io.uhndata.cards.versioning:staleFormsCheckin=[autocheckin]"
+            ]
+        }
+    }
+}
