+++ modules/versioning/src/main/java/io/uhndata/cards/versioning/StaleFormsCheckinTask.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.uhndata.cards.internal;
+package io.uhndata.cards.versioning;
 
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
